# openApi/swagger.yaml
openapi: 3.0.0
info:
  title: 'LaserLiga API'
  description: 'All API methods on the Laser liga portal.'
  version: '1.0'
servers:
  -
    url: 'https://laserliga.cz'
paths:
...
  /api/games:
    get:
      tags:
        - Games
      summary: 'Get list of all games'
      operationId: listGames
      parameters:
        -
          name: date
          in: query
          description: 'Filter games by date'
          required: false
          schema:
            type: string
            format: date
        -
          name: system
          in: query
          description: 'Filter games by system'
          required: false
          schema:
            type: string
        -
          name: returnLink
          in: query
          description: 'If specified, only game links will be returned'
          required: false
          schema:
            type: boolean
        -
          name: returnCodes
          in: query
          description: 'If specified, only game codes will be returned'
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: 'Successful operation. List of games is returned'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamesListResponse'
        '400':
          description: 'Request error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Games
      summary: 'Import games data'
      description: 'This method imports games data.'
      operationId: importGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - system
                - games
              properties:
                system:
                  type: string
                games:
                  type: array
                  items: { oneOf: [{ $ref: '#/components/schemas/Evo5GameImport' }, { $ref: '#/components/schemas/Evo6GameImport' }] }
              type: object
      responses:
        '201':
          description: 'Successful import'
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - { properties: { values: { properties: { imported: { description: 'Number of games imported', type: integer } }, type: object, example: { imported: 1 } } }, type: object }
                  - { $ref: '#/components/schemas/SuccessResponse' }
              example:
                message: 'Games imported'
                values:
                  imported: 1
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'Game belongs to a different arena'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 'Server error during save operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/games/stats:
    get:
      tags:
        - Games
      summary: 'Get Game Stats'
      description: 'This method returns statistical information for games.'
      operationId: stats
      parameters:
        -
          name: date
          in: query
          description: 'Filter stats by this date'
          required: false
          schema:
            type: string
          example: '2023-04-01'
        -
          name: system
          in: query
          description: 'Filter stats by this game system'
          required: false
          schema:
            type: string
          example: GameSystem1
      responses:
        '200':
          description: 'Game statistics'
          content:
            application/json:
              schema:
                properties:
                  games: { type: integer }
                  players: { type: integer }
                  teams: { type: integer }
                type: object
...
  '/api/devtools/users/{id}/stats':
    get:
      tags:
        - Devtools
        - Users
      summary: 'Update User Stats'
      description: 'This method updates the stats for the specified user.'
      operationId: updateStats
      parameters:
        -
          name: id
          in: path
          description: 'User ID'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerStats'
  /api/devtools/users/stats:
    get:
      tags:
        - Devtools
        - Users
      summary: 'Update All User Stats'
      description: 'This method updates the stats for all users.'
      operationId: updateAllUsersStats
      parameters:
        -
          name: from
          in: query
          description: 'User id to start processing from'
          required: false
          schema:
            type: integer
          example: 0
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/PlayerStats'
...
components:
  schemas:
    GamesListResponse:
      type: array
      items:
        oneOf:
          -
            $ref: '#/components/schemas/Game'
          -
            type: string
          -
            type: string
...
    Game:
      description: 'Base class for game models'
      properties:
        resultsFile:
          type: string
          nullable: true
        modeName:
          type: string
        fileTime:
          type: string
          format: date-time
          nullable: true
        start:
          type: string
          format: date-time
          nullable: true
        importTime:
          type: string
          format: date-time
          nullable: true
        end:
          type: string
          format: date-time
          nullable: true
        timing:
          nullable: true
        code:
          type: string
        mode:
          oneOf:
            -
              $ref: '#/components/schemas/AbstractMode'
          nullable: true
          description: $mode
        gameType: {  }
        arena:
          oneOf:
            -
              $ref: '#/components/schemas/Arena'
          nullable: true
        music:
          oneOf:
            -
              $ref: '#/components/schemas/MusicMode'
          nullable: true
        group:
          description: ''
          nullable: true
        started:
          type: boolean
        finished:
          type: boolean
        visited:
          type: boolean
      type: object
    AbstractMode:
      description: 'Base class for all game mode models'
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type: {  }
        loadName:
          type: string
          nullable: true
        systems:
          type: string
          nullable: true
        teams:
          type: string
        settings:
          $ref: '#/components/schemas/ModeSettings'
        rankable:
          type: boolean
      type: object
...
    ModeSettings:
      description: 'Data-model for all game mode settings'
      properties:
        public:
          type: boolean
        mines:
          type: boolean
        partWin:
          type: boolean
        partTeams:
          type: boolean
        partPlayers:
          type: boolean
        partHits:
          type: boolean
        partBest:
          type: boolean
        partBestDay:
          type: boolean
        playerScore:
          type: boolean
        playerShots:
          type: boolean
        playerMiss:
          type: boolean
        playerAccuracy:
          type: boolean
        playerMines:
          type: boolean
        playerPlayers:
          type: boolean
        playerPlayersTeams:
          type: boolean
        playerKd:
          type: boolean
        playerFavourites:
          type: boolean
        playerLives:
          type: boolean
        teamScore:
          type: boolean
        teamAccuracy:
          type: boolean
        teamShots:
          type: boolean
        teamHits:
          type: boolean
        teamZakladny:
          type: boolean
        bestScore:
          type: boolean
        bestHits:
          type: boolean
        bestDeaths:
          type: boolean
        bestAccuracy:
          type: boolean
        bestHitsOwn:
          type: boolean
        bestDeathsOwn:
          type: boolean
        bestShots:
          type: boolean
        bestMiss:
          type: boolean
        bestMines:
          type: boolean
      type: object
...
    Arena:
      properties:
        name:
          type: string
          example: 'Laser arena Písek'
        gameCodePrefix:
          type: string
          example: g
        lat:
          type: number
          format: float
          example: 49.307678
          nullable: true
        lng:
          type: number
          format: float
          example: 14.147773
          nullable: true
        address:
          $ref: '#/components/schemas/Address'
        web:
          type: string
          example: 'https://laserarenapisek.cz'
          nullable: true
        contactEmail:
          type: string
          example: info@laserarenapisek.cz
          nullable: true
        contactPhone:
          type: string
          example: '+420 776 606 631'
          nullable: true
      type: object
...
    Evo5GameImport:
      type: object
      allOf:
        -
          $ref: '#/components/schemas/GameImport'
        -
          properties:
            blastShots:
              type: boolean
            reloadClips:
              type: integer
            allowFriendlyFire:
              type: boolean
            antiStalking:
              type: boolean
            scoring:
              nullable: true
            players:
              type: array
              items:
                $ref: '#/components/schemas/Evo5PlayerImport'
          type: object
    Evo5PlayerImport:
      type: object
      allOf:
        -
          $ref: '#/components/schemas/PlayerImport'
        -
          properties:
            scoreAccuracy:
              type: integer
              example: 128
              nullable: true
            scoreVip:
              type: integer
              example: 500
              nullable: true
            minesHits:
              type: integer
              example: 2
              nullable: true
            bonus:
              nullable: true
          type: object
    Evo6GameImport:
      type: object
      allOf:
        -
          $ref: '#/components/schemas/GameImport'
        -
          properties:
            blastShots:
              type: boolean
            reloadClips:
              type: integer
            allowFriendlyFire:
              type: boolean
            antiStalking:
              type: boolean
            scoring:
              nullable: true
            players:
              type: array
              items:
                $ref: '#/components/schemas/Evo6PlayerImport'
            triggerSpeed: {  }
            gameStyleType: {  }
            vipSettings:
              nullable: true
            zombieSettings:
              nullable: true
            hitGainSettings:
              nullable: true
            respawnSettings:
              nullable: true
          type: object
    Evo6PlayerImport:
      type: object
      allOf:
        -
          $ref: '#/components/schemas/PlayerImport'
        -
          properties:
            scoreAccuracy:
              type: integer
              example: 128
              nullable: true
            scoreVip:
              type: integer
              example: 500
              nullable: true
            minesHits:
              type: integer
              example: 2
              nullable: true
            bonuses:
              type: integer
              example: 10
              nullable: true
            calories:
              type: integer
              example: 200
              nullable: true
            activity:
              type: integer
              example: 200
              nullable: true
            penaltyCount:
              type: integer
              example: 5
              nullable: true
            scorePenalty:
              type: integer
              example: -500
              nullable: true
            scoreEncouragement:
              type: integer
              example: 1000
              nullable: true
            scoreActivity:
              type: integer
              example: 1000
              nullable: true
            scoreKnockout:
              type: integer
              example: 1000
              nullable: true
            scoreReality:
              type: integer
              example: 1000
              nullable: true
          type: object
    GameImport:
      properties:
        gameType:
          nullable: true
        lives:
          type: integer
          example: 999
          nullable: true
        ammo:
          type: integer
          example: 9999
          nullable: true
        modeName:
          type: string
          example: 'Team deathmach'
          nullable: true
        fileNumber:
          example: '0123'
          nullable: true
        code:
          type: string
          example: g6739f7146ded5
          nullable: true
        respawn:
          type: integer
          example: 5
          nullable: true
        sync:
          type: boolean
          example: true
          nullable: true
        fileTime:
          type: string
          format: date-time
        importTime:
          type: string
          format: date-time
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        timing:
          nullable: true
        mode:
          oneOf:
            -
              $ref: '#/components/schemas/ModeImport'
          nullable: true
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerImport'
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamImport'
        music:
          oneOf:
            -
              $ref: '#/components/schemas/MusicImport'
          nullable: true
        group:
          oneOf:
            -
              $ref: '#/components/schemas/GroupImport'
          nullable: true
      type: object
    GroupImport:
      properties:
        id:
          description: 'Local group ID'
          type: integer
          example: 1
        name:
          type: string
          example: 'Skupina ABC'
      type: object
    ModeImport:
      properties:
        type:
          nullable: true
        name:
          type: string
          example: 'Team deathmach'
        description:
          type: string
        settings:
          oneOf:
            -
              $ref: '#/components/schemas/ModeSettings'
          nullable: true
        rankable:
          type: boolean
        order:
          type: integer
      type: object
    MusicImport:
      properties:
        id:
          description: 'Local music mode ID'
          type: integer
          example: 1
      type: object
...
    PlayerImport:
      properties:
        id:
          type: integer
          example: 1
          nullable: true
        id_player:
          type: integer
          example: 1
          nullable: true
        name:
          type: string
          example: Heroyt
          nullable: true
        code:
          type: string
          example: 1-8HRT8
          nullable: true
        team:
          nullable: true
          oneOf:
            -
              description: 'Team color'
              type: integer
              example: 1
            -
              $ref: '#/components/schemas/TeamColorImport'
        score:
          type: integer
          example: 10000
          nullable: true
        skill:
          type: integer
          example: 1000
          nullable: true
        shots:
          type: integer
          example: 100
          nullable: true
        accuracy:
          type: integer
          example: 50
          nullable: true
        vest:
          type: integer
          example: 1
          nullable: true
        hits:
          type: integer
          example: 100
          nullable: true
        deaths:
          type: integer
          example: 50
          nullable: true
        hitsOwn:
          type: integer
          example: 0
          nullable: true
        deathsOwn:
          type: integer
          example: 0
          nullable: true
        hitsOther:
          type: integer
          example: 100
          nullable: true
        deathsOther:
          type: integer
          example: 50
          nullable: true
        vip:
          type: boolean
          example: false
          nullable: true
        birthday:
          type: boolean
          example: false
          nullable: true
        myLasermaxx:
          type: string
          format: url
          example: ''
          nullable: true
        hitPlayers:
          type: array
          items:
            $ref: '#/components/schemas/PlayerHitImport'
        position:
          type: integer
          example: 1
          nullable: true
        shotPoints:
          type: integer
          example: 0
          nullable: true
        scoreBonus:
          type: integer
          example: 0
          nullable: true
        scoreMines:
          type: integer
          example: -50
          nullable: true
        ammoRest:
          type: integer
          example: 9899
          nullable: true
        tournamentPlayer:
          type: integer
          example: 1
          nullable: true
      type: object
    TeamColorImport:
      properties:
        id:
          type: integer
          example: 1
          nullable: true
        color:
          type: integer
          example: 1
      type: object
    TeamImport:
      properties:
        id:
          type: integer
          example: 1
          nullable: true
        id_team:
          type: integer
          example: 1
          nullable: true
        name:
          type: string
          example: 'Modrý tým'
          nullable: true
        score:
          type: integer
          example: 10000
          nullable: true
        color:
          type: integer
          example: 1
          nullable: true
        position:
          type: integer
          example: 1
          nullable: true
        tournamentTeam:
          type: integer
          example: 1
          nullable: true
        bonus:
          type: integer
          example: 500
          nullable: true
      type: object
...
    PlayerStats:
      properties:
        gamesPlayed:
          type: integer
        arenasPlayed:
          type: integer
        rank:
          type: integer
        averageAccuracy:
          type: number
          format: float
        averagePosition:
          type: number
          format: float
        maxAccuracy:
          type: integer
        maxScore:
          type: integer
        maxSkill:
          type: integer
        shots:
          type: integer
        averageShots:
          type: number
          format: float
        averageShotsPerMinute:
          type: number
          format: float
        totalMinutes:
          type: integer
        kd:
          type: number
          format: float
        hits:
          type: integer
        deaths:
          type: integer
      type: object
...
    MusicMode:
      properties:
        name:
          type: string
        order:
          type: integer
        fileName:
          type: string
        arena:
          oneOf:
            -
              $ref: '#/components/schemas/Arena'
          nullable: true
        idLocal:
          type: integer
        previewStart:
          type: integer
      type: object
...
    ErrorResponse:
      properties:
        title:
          type: string
          example: 'Error title'
        type:
          $ref: '#/components/schemas/ErrorType'
        detail:
          type: string
          example: 'Error description'
          nullable: true
        exception:
          properties:
            message:
              type: string
              example: 'Some exception description'
            code:
              type: integer
              example: 123
            trace:
              type: array
              items:
                type: object
              example:
                -
                  file: index.php
                  line: 1
                  function: abc
                  args: ['Argument value']
          type: object
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
    SuccessResponse:
      properties:
        message:
          type: string
          example: Message
        detail:
          type: string
          example: Description
          nullable: true
        values:
          type: object
          example:
            key1: value1
            key2: value2
          nullable: true
      type: object
    ErrorType:
      type: string
      enum:
        - validation_error
        - database_error
        - internal_error
        - resource_not_found_error
        - resource_access_error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  -
    bearerAuth: []
...